requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
ell-ai
openai
anthropic
flask-cors
flask-jwt-extended
gunicorn
python-magic
validators
tiktoken
tenacity
pyyaml
markdown
beautifulsoup4
Werkzeug
Jinja2
itsdangerous
click
blinker
cachelib